IdeaInspiration/Sources/Community/QASource/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "prismq-idea-sources-community-qasource"
version = "1.0.0"
description = "PrismQ module for gathering questions from Q&A platforms"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "PrismQ", email = "dev@prismq.com"}
]
keywords = ["prismq", "ai", "qa", "stackoverflow", "stackexchange"]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "sqlalchemy>=2.0.0",
    "vaderSentiment>=3.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
qa-source = "src.cli:main"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["_meta/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"
